Crate tiralabra

Modules

math
Defines mathematical structures and operations, primarily relating to complex numbers.

ring_buffer
Ring buffer for transferring audio data between threads in a wait-free manner.

util
Miscellaneous array utilities.

Structs

CorrelationMatch
Finds the closest match of a shorter piece of audio from a larger piece of audio.

CrossCorrelation
Computes cross correlation efficiently, using FFT.

DisplayBuffer
Stores a prepared CorrelationMatch and buffers for display and memory.

Fft
Implements the FFT, i.e. Fast Fourier Transform, and its inverse.

Functions

parabolic_interpolation_minimum
Finds the approximate minimum point of a function given three fixed points, using a parabola.