Module tiralabra::correlation_match::parabolic_interpolation

Finds the approximate minimum point of a function given three fixed points, using a parabola.

Formulas derived in /dokumentaatio/parabolic_interpolation_formulas.py

Constants

EPS

Functions

get_minimum_point
Interpolate a function given three points, (0, a), (1, b), (2, c), and find the approximate minimum point. Returns a pair (x, y), describing the minimum point. May also return None, if there is no minimum point or it is not on the interval [0, 2].