Function tiralabra::ring_buffer::with_capacity

pub fn with_capacity<T: Item>(capacity: usize) -> (Producer<T>, Consumer<T>)

Construct a ring buffer with a given capacity, and return (producer, consumer). The producer can then be sent to another thread for example.

Internally, an allocation that is one larger than the given capacity is made. This is because for an allocation of size capacity + 1, there are exactly capacity + 1 different offsets the read and write indices can point to, which corresponds to the range 0..=capacity.