Trait tiralabra::ring_buffer::Item

pub trait Item: Copy + Send + Sync + Default { }

Defines the interface that an item stored in the queue must implement. Only very simple types, such as floating point numbers or small structs that implement Copy are permitted.