Struct tiralabra::DisplayBuffer

pub struct DisplayBuffer { /* fields omitted */ }

Stores a prepared CorrelationMatch and buffers for display and memory.

Implementations

impl DisplayBuffer

pub fn new(input_size: usize, display_size: usize) -> Self

Construct a new DisplayBuffer with given input buffer size and display buffer size.

input_size must be at least as learge as display_size.

The weight function is populated with a Hann window, so the center of the display is prioritized when matching.

pub fn scroll(&mut self, amount: i32)

Scroll all internal buffers by the given signed amount of samples, to the right.

Missing data is retrieved from the input buffer, or replaced with zeros if not available.

pub fn get_buffer_mut(&mut self) -> &mut [Num]

Get a mutable reference to the input buffer. If it is mutated, remember to call update_match afterwards.

The length of the slice is input_size given on construction.

pub fn update_match(
&mut self,
stabilize: bool,
memory_decay: Num,
period_decay: Num
)

Update the correlation match position, memory buffer and period estimate based on newest data.

If stabilize is set to false, no matching is performed, and the previously set offset is retained.

The memory_decay and period_decay parameters determine the decay coefficient of the memory buffer, and the interval average respectively:

average = coeff * new + (1. - coeff) * average;

Set to 1.0 to bypass smoothing.

update_display should be called separately to update the display buffer.

pub fn update_display(&mut self, display_decay: Num)

Update the display buffer based on the newest input data and matched offset.

This method may be called more often than update_match, even when there is no new data, to animate smoothly.

pub fn get_display(&self) -> &[Num]

Retrieve the contents of the display buffer.

The length of the slice is display_size given on construction.

pub fn get_memory(&self) -> &[Num]

Retrieve the contents of the memory buffer. This is what is used to find a match in update_match.

The length of the slice is display_size given on construction.

pub fn get_period(&self) -> Num

Get current estimated period.

The fundamental frequency may be obtained via the sampling rate as follows:

f = SAMPLE_RATE / period

pub fn get_offset(&self) -> (usize, Num)

Get the current offset and residual.

The first item of the tuple is a whole number, denoting the starting index of the latest match in the input buffer. The second item denotes accumulated subsample precision, which is less than 1 by absolute value. A plot of the waveform should be offset by the negation of the residual.