Module tiralabra::util

Miscellaneous array utilities.

Structs

IterWindows
Iterate over fixed size windows.

Functions

shift_left
Shift the contents of the given array left, i.e. towards negative indices, and replace the empty positions with the contents of replace. The amount to be shifted left is determined by replace.len().

shift_left_fill
Like shift_left, but shifts by a given amount and fills outside with the value set.

shift_right
Similar to shift_left, but in the opposite direction. The contents of replace are placed in the beginning of array.

shift_right_fill
Like shift_right, but shifts by a given amount and fills outside with the value set.